@@ -2,7 +2,6 @@
   "name": "anchor",
   "version": "1.0.0",
   "description": "A user system API starter",
-  "main": "index.js",
   "scripts": {
     "start": "nodemon --inspect server.js",
     "first-time-setup": "node first-time-setup.js",
@@ -22,56 +21,55 @@
   },
   "homepage": "https://github.com/hicsail/anchor#readme",
   "engines": {
-    "node": "8.x.x"
+    "node": "18.x.x"
   },
   "dependencies": {
+    "@hapi/basic": "^6.0.0",
+    "@hapi/cookie": "^11.0.2",
+    "@hapi/glue": "^8.0.0",
+    "@hapi/hapi": "^20.2.2",
+    "@hapi/hoek": "^10.0.0",
+    "@hapi/inert": "^6.0.5",
     "@hapi/joi": "^17.1.1",
-    "adm-zip": "0.4.x",
-    "archiver": "2.x.x",
-    "async": "2.x.x",
-    "bcrypt": "3.x.x",
+    "@hapi/vision": "^6.1.0",
+    "adm-zip": "0.5.x",
+    "archiver": "5.x.x",
+    "async": "3.x.x",
+    "bcrypt": "5.x.x",
     "boom": "5.x.x",
     "child_process": "1.x.x",
-    "confidence": "3.x.x",
-    "dotenv": "4.x.x",
-    "glue": "5.x.x",
-    "handlebars": "4.x.x",
-    "hapi": "17.x.x",
-    "hapi-auth-basic": "5.x.x",
-    "hapi-auth-cookie": "9.x.x",
-    "hapi-auth-jwt2": "8.x.x",
-    "hapi-cron": "1.x.x",
-    "hapi-react-views": "^10.0.0",
-    "hapi-remote-address": "1.x.x",
-    "hapi-swagger": "9.x.x",
-    "hoek": "5.x.x",
-    "inert": "^5.1.0",
-    "joi": "13.x.x",
+    "confidence": "5.x.x",
+    "dotenv": "16.x.x",
+    "hapi-auth-jwt2": "^10.2.0",
+    "hapi-cron": "^1.1.0",
+    "hapi-remote-address": "^1.0.0",
+    "hapi-swagger": "^14.5.5",
+    "joi": "^17.6.0",
     "joi-json": "^4.0.0",
-    "joi-password-complexity": "2.x.x",
+    "joi-password-complexity": "5.x.x",
     "jsonwebtoken": "8.x.x",
-    "lout": "10.x.x",
-    "mongo-models": "1.x.x",
-    "mongodb": "2.x.x",
-    "nodemailer": "4.x.x",
+    "lout": "11.x.x",
+    "mongo-models": "3.x.x",
+    "mongodb": "3.x.x",
+    "nodemailer": "6.x.x",
     "nodemailer-markdown": "1.x.x",
     "object-assign": "4.x.x",
-    "slug": "0.9.x",
+    "slug": "5.3.x",
     "useragent": "2.x.x",
-    "uuid": "3.x.x",
-    "vision": "5.3.3",
-    "visionary": "6.x.x"
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
+    "@babel/eslint-parser": "^7.18.2",
+    "@hapi/eslint-plugin": "^6.0.0",
+    "@hapi/lab": "^25.0.1",
     "code": "5.x.x",
     "coveralls": "3.x.x",
-    "lab": "15.x.x",
-    "nodemon": "1.x.x",
-    "npm-run": "^4.1.2",
+    "nodemon": "2.x.x",
+    "npm-run": "^5.0.1",
     "pre-commit": "1.x.x",
     "prompt-sync": "^4.2.0",
-    "promptly": "2.x.x",
-    "proxyquire": "1.x.x",
+    "promptly": "3.x.x",
+    "proxyquire": "2.x.x",
     "yargs": "^17.5.1"
   },
   "pre-commit": [
