@@ -25,7 +25,7 @@ const register = function (server, options) {
         userId = request.auth.credentials.user._id.toString();
       }
 
-      const event = await Event.create(request.params.name, userId);
+      const event = await Event.create({ name: request.params.name, userId: userId });
 
       return event;
     }
