@@ -3,7 +3,7 @@ const Feedback = require('../../../server/models/feedback');
 const Code = require('code');
 const Config = require('../../../config');
 const Fixtures = require('../fixtures');
-const Lab = require('lab');
+const Lab = require('@hapi/lab');
 
 const lab = exports.lab = Lab.script();
 const config = Config.get('/hapiAnchorModel/mongodb');
@@ -29,7 +29,7 @@ lab.experiment('Feedback Model', () => {
     const description = 'description';
     const userId = 'testId';
 
-    const feedback = await Feedback.create(subject, description, userId);
+    const feedback = await Feedback.create({ subject, description, userId });
 
     Code.expect(feedback).to.be.an.instanceOf(Feedback);
     Code.expect(feedback.subject).to.equal('subject');
